KAN-216 <message: 스트리밍 방식으로 다운로드>
