+++ build.gradle
@@ -29,6 +29,7 @@ dependencies {
 	implementation 'org.apache.tika:tika-core:2.7.0'
 	implementation 'com.squareup.okhttp3:okhttp'
 	implementation 'org.modelmapper:modelmapper:3.0.0'
+	implementation 'software.amazon.awssdk:s3:2.20.0'
 	compileOnly 'org.projectlombok:lombok'
 	developmentOnly 'org.springframework.boot:spring-boot-devtools'
 	runtimeOnly 'com.h2database:h2'
