+++ src/main/java/com/GASB/anti_malware/config/S3Config.java
@@ -1,0 +1,17 @@
+package com.GASB.anti_malware.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import software.amazon.awssdk.regions.Region;
+import software.amazon.awssdk.services.s3.S3Client;
+
+@Configuration
+public class S3Config {
+
+//    @Bean
+//    public S3Client s3Client() {
+//        return S3Client.builder()
+//                .region(Region.of("your-region")) // 원하는 AWS 리전을 설정합니다.
+//                .build();
+//    }
+}
