+++ src/main/java/com/GASB/anti_malware/controller/VirusTotalController.java
@@ -20,17 +20,17 @@ public class VirusTotalController {
 
     private final StoredFileService storedFileService;
     private final VtReportSaveService vtReportSaveService;
-    private final VtUploadService vtUploadService;
+    private final TempVtUploadService tempVtUploadService;
     private final VtAnalysisService vtAnalysisService;
     private final FileStatusService fileStatusService;
 
     @Autowired
     public VirusTotalController(StoredFileService storedFileService,
-                                VtReportSaveService vtReportSaveService, VtUploadService vtUploadService,
+                                VtReportSaveService vtReportSaveService, TempVtUploadService tempVtUploadService,
                                 VtAnalysisService vtAnalysisService, FileStatusService fileStatusService) {
         this.storedFileService = storedFileService;
         this.vtReportSaveService = vtReportSaveService;
-        this.vtUploadService = vtUploadService;
+        this.tempVtUploadService = tempVtUploadService;
         this.vtAnalysisService = vtAnalysisService;
         this.fileStatusService = fileStatusService;
     }
@@ -54,13 +54,13 @@ public class VirusTotalController {
             Map<Long, String> results = new HashMap<>();
             for (Long fileId : vtRequestDto.getFileIds()) {
                 // 파일 상태 확인
-                int vtStatus = storedFileService.getVtStatusByFileId(fileId);
+                int vtStatus = fileStatusService.getVtStatusByFileId(fileId);
                 if (vtStatus != -1) {
                     results.put(fileId, "Report Already exists.");
                     continue;
                 }
 
-                Result result = vtUploadService.uploadFile(fileId); // 파일 ID로 업로드 처리
+                Result result = tempVtUploadService.uploadFile(fileId); // 파일 ID로 업로드 처리
                 results.put(fileId, result.getMessage());
 
                 if (result.isSuccess()) {
