+++ src/main/java/com/GASB/anti_malware/service/FileStatusService.java
@@ -18,7 +18,7 @@ import java.util.stream.Collectors;
 @Service
 public class FileStatusService {
 
-    private static final Logger logger = LoggerFactory.getLogger(VtUploadService.class);
+    private static final Logger logger = LoggerFactory.getLogger(TempVtUploadService.class);
 
     private final FileStatusRepository fileStatusRepository;
     private final ModelMapper modelMapper;
@@ -69,5 +69,10 @@ public class FileStatusService {
 
     private FileStatusDto convertToFileStatusDto(FileStatus fileStatus) {
         return modelMapper.map(fileStatus, FileStatusDto.class);
+    }
+
+    public int getVtStatusByFileId(Long fileId) {
+        FileStatus fileStatus = fileStatusRepository.findByStoredFileId(fileId);
+        return fileStatus != null ? fileStatus.getVtStatus() : -1; // 상태가 없으면 -1 반환
     }
 }
