+++ src/main/java/com/GASB/anti_malware/service/StoredFileService.java
@@ -22,14 +22,12 @@ import java.util.stream.Collectors;
 public class StoredFileService {
 
         private final StoredFileRepository storedFileRepository;
-        private final FileStatusRepository fileStatusRepository;
         private final ModelMapper modelMapper;
 
         @Autowired
-        public StoredFileService(StoredFileRepository storedFileRepository, FileStatusRepository fileStatusRepository
-        ,ModelMapper modelMapper) {
+        public StoredFileService(StoredFileRepository storedFileRepository
+                , ModelMapper modelMapper) {
             this.storedFileRepository = storedFileRepository;
-            this.fileStatusRepository = fileStatusRepository;
             this.modelMapper = modelMapper;
         }
 
@@ -46,53 +44,49 @@ public class StoredFileService {
             return storedFileRepository.findBySaltedHash(saltedHash).isPresent();
         }
 
-        public int getVtStatusByFileId(Long fileId) {
-            FileStatus fileStatus = fileStatusRepository.findByStoredFileId(fileId);
-            return fileStatus != null ? fileStatus.getVtStatus() : -1; // 상태가 없으면 -1 반환
+        // 전체 파일 리스트 반환
+        public ResponseDto.FileListResponse getFileList() {
+            List<FileListDto> fileList = storedFileRepository.findAll()
+                    .stream()
+                    .map(storedFile -> {
+                        VtReport vtReport = storedFile.getVtReport();
+                        FileStatus fileStatus = storedFile.getFileStatus();
+                        ScanTable scanTable = storedFile.getScanTable();
+                        VtReportDto vtReportDto = (vtReport != null) ? convertToVtReportDto(vtReport) : null;
+                        FileStatusDto fileStatusDto = convertToFileStatusDto(fileStatus);
+
+                        InnerScanDto innerScanDto = InnerScanDto.builder()
+                                .step2Detail(null) // Set step2Detail as needed
+                                .build();
+
+                        return FileListDto.builder()
+                                .id(storedFile.getId())
+                                .saltedHash(storedFile.getSaltedHash())
+                                .size(storedFile.getSize())
+                                .type(storedFile.getType())
+                                .vtReport(vtReportDto)
+                                .fileStatus(fileStatusDto)
+                                .GScan(innerScanDto)
+                                .build();
+                    }).collect(Collectors.toList());
+
+            int totalFiles = fileList.size();
+            int malwareTotal = (int) fileList.stream()
+                    .filter(fileListDto -> {
+                        VtReportDto vtReport = fileListDto.getVtReport();
+                        return vtReport != null && !vtReport.getThreatLabel().equals("none");
+                    })
+                    .count();
+
+            return ResponseDto.FileListResponse.of(totalFiles, malwareTotal, fileList);
         }
 
-    public ResponseDto.FileListResponse getFileList() {
-        List<FileListDto> fileList = storedFileRepository.findAll()
-                .stream()
-                .map(storedFile -> {
-                    VtReport vtReport = storedFile.getVtReport();
-                    FileStatus fileStatus = storedFile.getFileStatus();
-                    ScanTable scanTable = storedFile.getScanTable();
-                    VtReportDto vtReportDto = (vtReport != null) ? convertToVtReportDto(vtReport) : null;
-                    FileStatusDto fileStatusDto = convertToFileStatusDto(fileStatus);
+        private VtReportDto convertToVtReportDto(VtReport vtReport) {
+                return modelMapper.map(vtReport, VtReportDto.class);
+        }
 
-                    InnerScanDto innerScanDto = InnerScanDto.builder()
-                            .step2Detail(null) // Set step2Detail as needed
-                            .build();
-
-                    return FileListDto.builder()
-                            .id(storedFile.getId())
-                            .saltedHash(storedFile.getSaltedHash())
-                            .size(storedFile.getSize())
-                            .type(storedFile.getType())
-                            .vtReport(vtReportDto)
-                            .fileStatus(fileStatusDto)
-                            .GScan(innerScanDto)
-                            .build();
-                }).collect(Collectors.toList());
-
-        int totalFiles = fileList.size();
-        int malwareTotal = (int) fileList.stream()
-                .filter(fileListDto -> {
-                    VtReportDto vtReport = fileListDto.getVtReport();
-                    return vtReport != null && !vtReport.getThreatLabel().equals("none");
-                })
-                .count();
-
-        return ResponseDto.FileListResponse.of(totalFiles, malwareTotal, fileList);
-    }
-
-    private VtReportDto convertToVtReportDto(VtReport vtReport) {
-            return modelMapper.map(vtReport, VtReportDto.class);
-    }
-
-    private FileStatusDto convertToFileStatusDto(FileStatus fileStatus){
-            return modelMapper.map(fileStatus, FileStatusDto.class);
-    }
+        private FileStatusDto convertToFileStatusDto(FileStatus fileStatus){
+                return modelMapper.map(fileStatus, FileStatusDto.class);
+        }
 }
