+++ src/main/java/com/GASB/anti_malware/service/VtReportSaveService.java
@@ -38,18 +38,16 @@ public class VtReportSaveService {
     private final ModelMapper modelMapper;
     private final VtReportRepository reportRepository;
     private final StoredFileRepository storedFileRepository;
-    private final StoredFileService storedFileService;
     private final FileStatusService fileStatusService;
 
     @Autowired
     public VtReportSaveService(RestTemplate restTemplate, VtReportRepository reportRepository, StoredFileRepository storedFileRepository
-            , ModelMapper modelMapper, StoredFileService storedFileService, FileStatusService fileStatusService) {
+            , ModelMapper modelMapper, FileStatusService fileStatusService) {
         this.restTemplate = restTemplate;
         this.objectMapper = new ObjectMapper();
         this.reportRepository = reportRepository;
         this.storedFileRepository = storedFileRepository;
         this.modelMapper = modelMapper;
-        this.storedFileService = storedFileService;
         this.fileStatusService = fileStatusService;
     }
