+++ src/test/java/com/GASB/anti_malware/service/VtReportSaveServiceTest.java
@@ -42,13 +42,13 @@ public class VtReportSaveServiceTest {
 
         @InjectMocks
         private VtReportSaveService vtReportSaveService;
-
+        private FileStatusService fileStatusService;
         private StoredFileService storedFileService;
         private ObjectMapper objectMapper = new ObjectMapper();
 
         @BeforeEach
         void setUp() {
-            vtReportSaveService = new VtReportSaveService(restTemplate, reportRepository, storedFileRepository, modelMapper,storedFileService);
+            vtReportSaveService = new VtReportSaveService(restTemplate, reportRepository, storedFileRepository, modelMapper,fileStatusService);
         }
 
         @Test
